LinkedIn_AI_AGENT/requirements.txt
requests>=2.29.0
beautifulsoup4>=4.12.2
cryptography>=41.0.3
schedule>=1.2.0
pytz>=2024.1
python-dotenv>=1.0.0